requirements.txt
fastapi
uvicorn
python-telegram-bot
schedule
mcp-python-sdk 